refactor: Move proxy components into crate::net
