+++ src/net/io/completion.rs
@@ -1,0 +1,2 @@
+#[cfg(target_os = "linux")]
+pub mod io_uring;
