+++ src/net/io/nic.rs
@@ -1,0 +1,2 @@
+#[cfg(target_os = "linux")]
+pub mod xdp;
