+++ src/net/io/poll/tokio.rs
@@ -16,7 +16,35 @@
 
 //! The reference implementation is used for non-Linux targets
 
-impl super::DownstreamReceiveWorkerConfig {
+/// On linux spawns a io-uring runtime + thread, everywhere else spawns a regular tokio task.
+macro_rules! uring_spawn {
+    ($span:expr_2021, $future:expr_2021) => {{
+        let (tx, rx) = std::sync::mpsc::channel::<()>();
+        use tracing::Instrument as _;
+
+        use tracing::instrument::WithSubscriber as _;
+
+        let fut = async move {
+            let _ = tx.send(());
+            $future.await
+        };
+
+        if let Some(span) = $span {
+            tokio::spawn(fut.instrument(span).with_current_subscriber());
+        } else {
+            tokio::spawn(fut.with_current_subscriber());
+        }
+        rx
+    }};
+}
+
+macro_rules! uring_inner_spawn {
+    ($future:expr_2021) => {
+        tokio::spawn($future);
+    };
+}
+
+impl crate::net::io::Listener {
     pub fn spawn(self, packet_queue: crate::net::PacketQueue) -> eyre::Result<()> {
         let Self {
             worker_id,
@@ -109,7 +137,7 @@ impl super::DownstreamReceiveWorkerConfig {
                         match result {
                             Ok((_size, mut source)) => {
                                 source.set_ip(source.ip().to_canonical());
-                                let packet = super::DownstreamPacket { contents: buffer, source };
+                                let packet = crate::net::packet::DownstreamPacket { contents: buffer, source };
 
                                 if let Some(last_received_at) = last_received_at {
                                     crate::metrics::packet_jitter(
@@ -144,6 +172,102 @@ impl super::DownstreamReceiveWorkerConfig {
 
         use eyre::WrapErr as _;
         worker.recv().context("failed to spawn receiver task")?;
+        Ok(())
+    }
+}
+
+impl crate::net::sessions::SessionPool {
+    pub fn spawn_session(
+        self: std::sync::Arc<Self>,
+        raw_socket: socket2::Socket,
+        port: u16,
+        pending_sends: crate::net::PacketQueue,
+    ) -> Result<(), crate::net::error::PipelineError> {
+        let pool = self;
+
+        uring_spawn!(
+            uring_span!(tracing::debug_span!("session pool")),
+            async move {
+                let mut last_received_at = None;
+
+                let socket =
+                    std::sync::Arc::new(crate::net::DualStackLocalSocket::from_raw(raw_socket));
+                let socket2 = socket.clone();
+                let (tx, mut rx) = tokio::sync::oneshot::channel();
+
+                uring_inner_spawn!(async move {
+                    let (pending_sends, mut sends_rx) = pending_sends;
+                    let mut sends_double_buffer = Vec::with_capacity(pending_sends.capacity());
+
+                    while sends_rx.changed().await.is_ok() {
+                        if !*sends_rx.borrow() {
+                            tracing::trace!("io loop shutdown requested");
+                            break;
+                        }
+
+                        sends_double_buffer = pending_sends.swap(sends_double_buffer);
+
+                        for packet in sends_double_buffer.drain(..sends_double_buffer.len()) {
+                            let destination = packet.destination.as_socket().unwrap();
+                            tracing::trace!(
+                                %destination,
+                                length = packet.data.len(),
+                                "sending packet upstream"
+                            );
+                            let (result, _) = socket2.send_to(packet.data, destination).await;
+                            let asn_info = packet.asn_info.as_ref().into();
+                            match result {
+                                Ok(size) => {
+                                    crate::metrics::packets_total(crate::metrics::READ, &asn_info)
+                                        .inc();
+                                    crate::metrics::bytes_total(crate::metrics::READ, &asn_info)
+                                        .inc_by(size as u64);
+                                }
+                                Err(error) => {
+                                    tracing::trace!(%error, "sending packet upstream failed");
+                                    let source = error.to_string();
+                                    crate::metrics::errors_total(
+                                        crate::metrics::READ,
+                                        &source,
+                                        &asn_info,
+                                    )
+                                    .inc();
+                                    crate::metrics::packets_dropped_total(
+                                        crate::metrics::READ,
+                                        &source,
+                                        &asn_info,
+                                    )
+                                    .inc();
+                                }
+                            }
+                        }
+                    }
+
+                    let _ = tx.send(());
+                });
+
+                loop {
+                    let buf = pool.buffer_pool.clone().alloc();
+                    tokio::select! {
+                        received = socket.recv_from(buf) => {
+                            let (result, buf) = received;
+                            match result {
+                                Err(error) => {
+                                    tracing::trace!(%error, "error receiving packet");
+                                    crate::metrics::errors_total(crate::metrics::WRITE, &error.to_string(), &crate::metrics::EMPTY).inc();
+                                },
+                                Ok((_size, recv_addr)) => pool.process_received_upstream_packet(buf, recv_addr, port, &mut last_received_at),
+                            }
+                        }
+                        _ = &mut rx => {
+                            tracing::debug!("Closing upstream socket loop, downstream closed");
+                            return;
+                        }
+                    }
+                }
+            }
+        );
+
         Ok(())
     }
 }
